pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dorothy"
version = "0.1.0"
description = "Newspaper of averages - balanced news aggregation"
requires-python = ">=3.11"

dependencies = [
    # RSS Parsing
    "feedparser>=6.0.10",

    # OpenSearch
    "opensearch-py>=2.4.0",

    # Clustering
    "hdbscan>=0.8.33",
    "scikit-learn>=1.3.0",

    # Data modeling
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Configuration
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # HTTP client
    "httpx>=0.25.0",

    # Scheduling
    "schedule>=1.2.0",

    # Utilities
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "python-dateutil>=2.8.2",

    # Web server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "types-PyYAML>=6.0.0",
    "types-python-dateutil>=2.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
